+++ src/main/java/org/phoenix/planet/configuration/security/CustomAuthenticationEntryPoint.java
@@ -11,7 +11,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.phoenix.planet.constant.AuthenticationError;
 import org.phoenix.planet.constant.TokenKey;
-import org.phoenix.planet.error.TokenException;
+import org.phoenix.planet.error.auth.TokenException;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
