+++ src/main/java/org/phoenix/planet/controller/AuthController.java
@@ -10,7 +10,7 @@ import org.phoenix.planet.constant.AuthenticationError;
 import org.phoenix.planet.constant.TokenKey;
 import org.phoenix.planet.dto.member.request.SignUpRequest;
 import org.phoenix.planet.provider.TokenProvider;
-import org.phoenix.planet.service.AuthService;
+import org.phoenix.planet.service.MemberService;
 import org.phoenix.planet.util.CookieUtil;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
@@ -26,7 +26,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class AuthController {
 
     private final TokenProvider tokenProvider;
-    private final AuthService authService;
+    private final MemberService memberService;
 
     // 향후엔 없앨게.. (cookie 유틸이 @component scan 해야 되지 않나)
     private final CookieProperties cookieProps;
@@ -42,7 +42,7 @@ public class AuthController {
         @LoginMemberId long loginMemberId
     ) {
 
-        authService.signUp(loginMemberId, request, profileImage);
+        memberService.signUp(loginMemberId, request, profileImage);
 
         return ResponseEntity.ok().build();
     }
