+++ src/main/java/org/phoenix/planet/error/GlobalExceptionHandler.java
@@ -2,6 +2,7 @@ package org.phoenix.planet.error;
 
 import java.util.HashMap;
 import java.util.Map;
+import org.phoenix.planet.error.auth.TokenException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
