+++ src/main/java/org/phoenix/planet/error/auth/AuthException.java
@@ -1,4 +1,4 @@
-package org.phoenix.planet.error;
+package org.phoenix.planet.error.auth;
 
 import lombok.Getter;
 import org.phoenix.planet.constant.AuthenticationError;
