+++ src/main/java/org/phoenix/planet/provider/TokenProvider.java
@@ -18,8 +18,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.phoenix.planet.constant.AuthenticationError;
 import org.phoenix.planet.constant.Role;
 import org.phoenix.planet.dto.auth.PrincipalDetails;
-import org.phoenix.planet.dto.member.Member;
-import org.phoenix.planet.error.TokenException;
+import org.phoenix.planet.dto.member.raw.Member;
+import org.phoenix.planet.error.auth.TokenException;
 import org.phoenix.planet.service.RefreshTokenService;
 import org.phoenix.planet.util.CookieUtil;
 import org.springframework.beans.factory.annotation.Value;
