+++ src/main/java/org/phoenix/planet/service/MemberService.java
@@ -3,7 +3,7 @@ package org.phoenix.planet.service;
 import org.phoenix.planet.dto.member.request.SignUpRequest;
 import org.springframework.web.multipart.MultipartFile;
 
-public interface AuthService {
+public interface MemberService {
 
     void signUp(long loginMemberId, SignUpRequest request, MultipartFile profileImage);
 }
